Experiment with SQLite transaction support
Attempt to acquire write lock immediately when beginning a transaction
and retry every 1ms if this failed for up to 5 seconds.

In new Rails 8 projects, this is the default behaviour.

Disable the default queuing behaviour from SQLite based on busy_timeout.
Its backoff strategy is not ideal for web applications.

https://fractaledmind.github.io/2024/07/19/sqlite-in-ruby-backoff-busy-handler-problems/